+++ src/optimagic/optimization/optimize_result.py
@@ -7,7 +7,7 @@ import pandas as pd
 
 from optimagic import deprecations
 from optimagic.logging.logger import LogReader
-from optimagic.optimization.internal_optimization_problem import History
+from optimagic.optimization.history import History
 from optimagic.shared.compat import pd_df_map
 from optimagic.typing import PyTree
 from optimagic.utilities import to_pickle
