+++ tests/optimagic/optimization/test_algorithm.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 from optimagic.exceptions import InvalidAlgoInfoError, InvalidAlgoOptionError
 from optimagic.optimization.algorithm import AlgoInfo, Algorithm, InternalOptimizeResult
-from optimagic.optimization.internal_optimization_problem import HistoryEntry
+from optimagic.optimization.history import HistoryEntry
 from optimagic.typing import (
     AggregationLevel,
     EvalTask,
